Module reloading code (reloads on HUP)
This is much simpler than the previous version, reusing the configure class. I
think the previous version failed because entries have to be deleted from %INC
before being reloaded.

Warnings are still printed saying functionname redefined. This can be fixed
with "no warnings 'redefine';" in each module, but it would be nice to not
require module writers to have to do that. There is probably a way to set that
when loading the module from Config.pm.

git-svn-id: https://labs.omniti.com/resmon/branches/resmon2@278 8c0face9-b7db-6ec6-c4b3-d5f7145c7d55