Allow run_command to take a list instead of a string.
When it does this, commands are not executed via the shell, but are executed
directly. (See the documentation for 'exec' for an explanation of this).
Existing modules that pass a string are unaffected.

The rationale for this change is for the Pgrep (and similar) modules, where it
was picking up the 'sh -c pgrep' process and including it in the process
count. Executing pgrep directly prevents this.

git-svn-id: https://labs.omniti.com/resmon/branches/resmon2@309 8c0face9-b7db-6ec6-c4b3-d5f7145c7d55