Fix for a crash when run_command tries to run a non-existent command
When a child process exits (as is the case in run_command), the DESTROY method
of Resmon::Status gets called, and lots of bad things happen (web server is
killed, shared memory gets released). This commit adds a guard to make sure we
only do these if it is the main process that exits.

git-svn-id: https://labs.omniti.com/resmon/branches/resmon2@315 8c0face9-b7db-6ec6-c4b3-d5f7145c7d55