Remove the sysv IPC shared memory code in favor of using files instead.
This fixes an issue with resmon not correctly cleaning up the shared memory
segments when it dies.

Currently no locking is done on the file to prevent reading of a partial
status dump. This needs fixing.

Also, the file is deleted immediately after opening. This deals with the issue
of having a stale file around if resmon dies and also limits the ability of
somebody to inject invalid data into the file.

git-svn-id: https://labs.omniti.com/resmon/branches/resmon2@367 8c0face9-b7db-6ec6-c4b3-d5f7145c7d55