Support wildcard modules (fixes #7)
This is done using a wildcard_handler method in addition to the normal handler
method. The wildcard_handler method is called if * is in the config file as
the check name, and the handler module is called for normal checks. By
default, both methods will die with "Not implemented", meaning modules can
implement one or both of the methods, depending on whether they support normal
checks, wildcard checks, or some combination of both.

git-svn-id: https://labs.omniti.com/resmon/branches/resmon2@386 8c0face9-b7db-6ec6-c4b3-d5f7145c7d55