Greatly simplify how run_command works. (refs #14)
The original intent was to fix bad behavior when a command doesn't exist. Upon
looking further into the implementation of the run_command function, I
realized that using open achieved the same effect (backticks but we get a
process id out of it) without a lot of code that didn't quite work when the
exec failed.

git-svn-id: https://labs.omniti.com/resmon/branches/resmon2@414 8c0face9-b7db-6ec6-c4b3-d5f7145c7d55