Remove the stub Resmon::ExtComm and Resmon::Module modules (refs #12)
It's too much of a pain to have to change the built in versions of the modules
in the testmod script each time the resmon code changes, and it leads to odd
behavior with the testmod script that isn't present when run with resmon
itself.

The downside is that the testmod script is no longer standalone, and needs to
reference the resmon source code if it is moved out of the resmon tree. This
isn't as much of an issue as the testmod script by default is in the resources
directory of the resmon source and can be run direct from there.

git-svn-id: https://labs.omniti.com/resmon/branches/resmon2@415 8c0face9-b7db-6ec6-c4b3-d5f7145c7d55