Store the file handle as well as the pid in children. (fixes #14)
It is used in the clean up function. Incidentally, this fixes a really
annoying issue where the check timeout functionality broken. The cause of the
check timeout functionality breaking was a scoping issue - the reader file
handle falls out of scope when the alarm fires and dies, and it waits until
the process finishes before passing control back to resmon. Storing the file
handle prevents it from being cleaned up until after we kill the process in
the clean_up function later on.

git-svn-id: https://labs.omniti.com/resmon/branches/resmon2@416 8c0face9-b7db-6ec6-c4b3-d5f7145c7d55