support for HOSTS {ALLOW|DENY} lists
Hosts allow-deny lists are the coma or blank separated lists of
a dotted decimal IPv4 addresses of the form a.b.c.d. to match incoming machine’s IP address exactly,
or an 'ipaddr/n' where ipaddr is the IP address and n is the number of one bits in the netmask.
The first to occur match gives the result, skipping other rules.
If nothing matches IP is allowed.