Fix per-check interval/cached metrics
Credit: Michal Racek <michal.racek@t-systems.sk>

In Resmon 1, it was possible to add an interval parameter to an individual
check to give it a larger check interval than regular checks (cached results
would be returned in the meantime). This functionality wasn't properly carried
over into resmon 2, and this commit fixes that.